+++ ComputerVision/src/main/java/Converter/Convertion.java
@@ -14,13 +14,7 @@ public class Convertion implements I_ComputerVisionController {
     Util utility = new Util();
     I_Connection connection = new Darknet_Stub();
     Image img;
-    double upperHeight;
-    double lowerHeight;
-    double boxWidth;
-    double foundationBoxes[] = new double[4];
-    double drawPile[] = new double[1];
-    int numberOfBoxes = 7; //The number of boxes the width should be divided with.
-    double pileBoxes[] = new double[numberOfBoxes];
+
 
 
 
@@ -33,32 +27,7 @@ public class Convertion implements I_ComputerVisionController {
     }
 
 
-    public void calibrateImgBoxes(Image img){
 
-        lowerHeight = img.getHeight()/2;
-        upperHeight = img.getHeight()/2+lowerHeight;
-
-        //Creating boxessizes for piles
-        boxWidth = img.getWidth()/7;
-
-        //Creating gridboxes
-        for(int i = 1; i<=numberOfBoxes;i++){
-            pileBoxes[i-1]= boxWidth *i;
-            System.out.println("pileBox: "+pileBoxes[i-1]);
-
-            //Creating four foundation boxes
-            if(numberOfBoxes-i<=3){
-                foundationBoxes[numberOfBoxes-i] = boxWidth*i;
-                System.out.println("foundationBox: "+foundationBoxes[numberOfBoxes-i]);
-            }
-
-            //Creating bow for drawpile
-            if(numberOfBoxes-i==0){
-                drawPile[numberOfBoxes-i]=boxWidth*3;
-                System.out.println("drawPile: "+drawPile[numberOfBoxes-i]);
-            }
-        }
-    }
 
 
     private JsonArray ConvertImage(Image img){
